Rework notification mechanism when adding samples
Currently, we always notify the data users when any samples were added,
even when it was just one. As every such notification causes a repaint
of the affected trace(s), this is inefficient.

With this patch, we accumulate samples until a threshold is reached,
at which point we notify the users of the accumulated new samples.